Rect Greedy now uses CPython!
Holy moly guacamole, that was *not* fun to do, at all.
Some small changes are yet to be done, but still, not calling the greedy through console using files as intermediaries and instead calling it through a C library should be a huge improvement.
That being said, my sleep schedule is all sorts of messed up right now and I need to relax. Tomorrow I'll tidy this up.